+++ internal/updater/constants.go
@@ -3,5 +3,6 @@ package updater
 const (
 	allPackagesUpToDateMessage = "すべてのパッケージは最新です"
 	updateCommand              = "update"
+	upgradeCommand             = "upgrade"
 	latestVersionSuffix        = "@latest"
 )